[DEV-3543-3551-3549] change filters
